+++ crates/backend-uzu/src/encodable_block/weight_matrix.rs
@@ -118,18 +118,17 @@ impl<B: Backend> WeightMatrix<B> {
         let (rows, columns) = physical_shape(&layout, output_dim, input_dim);
 
         let Some(info) = quantization else {
-            let values =
-                tree.leaf("weights")?.validate(&[rows as u32, columns as u32], data_type)?.read_allocation()?;
+            let values = tree.leaf("weights")?.validate(&[rows, columns], data_type)?.read_allocation()?;
             return Ok(Self {
                 values,
                 quantized: None,
             });
         };
 
-        let group_size = info.group_size as usize;
+        let group_size = info.group_size;
         let packing_divisor = info.mode.packing_divisor();
         let storage_data_type = info.mode.storage_type();
-        if !columns.is_multiple_of(packing_divisor) {
+        if !columns.is_multiple_of(packing_divisor as u32) {
             return Err(WeightMatrixError::UnsupportedConfiguration(format!(
                 "stored columns {columns} are not divisible by packing divisor {packing_divisor}"
             )));
@@ -138,16 +137,16 @@ impl<B: Backend> WeightMatrix<B> {
 
         let values = tree
             .leaf("weights")?
-            .validate(&[rows as u32, (columns / packing_divisor) as u32], storage_data_type)?
+            .validate(&[rows, columns / packing_divisor as u32], storage_data_type)?
             .read_allocation()?;
-        let scales = tree.leaf("scales")?.validate(&[rows as u32, groups as u32], data_type)?.read_allocation()?;
+        let scales = tree.leaf("scales")?.validate(&[rows, groups], data_type)?.read_allocation()?;
         let correction = match info.method {
             QuantizationMethod::ScaleBias => QuantizedCorrection::Biases(
-                tree.leaf("biases")?.validate(&[rows as u32, groups as u32], data_type)?.read_allocation()?,
+                tree.leaf("biases")?.validate(&[rows, groups], data_type)?.read_allocation()?,
             ),
             QuantizationMethod::ScaleZeroPoint => QuantizedCorrection::ZeroPoints(
                 tree.leaf("zero_points")?
-                    .validate(&[rows as u32, groups.div_ceil(packing_divisor) as u32], storage_data_type)?
+                    .validate(&[rows, groups.div_ceil(packing_divisor as u32)], storage_data_type)?
                     .read_allocation()?,
             ),
             QuantizationMethod::ScaleSymmetric => QuantizedCorrection::Symmetric,
@@ -248,9 +247,9 @@ fn physical_shape(
     layout: &Layout,
     output_dim: u32,
     input_dim: u32,
-) -> (usize, usize) {
+) -> (u32, u32) {
     match layout {
-        Layout::OutputInput => (output_dim as usize, input_dim as usize),
-        Layout::InputOutput => (input_dim as usize, output_dim as usize),
+        Layout::OutputInput => (output_dim, input_dim),
+        Layout::InputOutput => (input_dim, output_dim),
     }
 }
